# Speed improements to point-source response computation from FDR