--- a/tests/response/test_full_detector_response.py
+++ b/tests/response/test_full_detector_response.py
@@ -52,6 +52,10 @@ def test_get_item():
 
         assert drm.unit.is_equivalent('m2')
 
+        drmu = response.get_pixel(0, weight=1.0 * u.s)
+
+        assert drmu.unit.is_equivalent('m2 s')
+                
         drm2 = response[0] # shorthand for get_pixel with unit weight
 
         assert drm == drm2