[draft PR to review coverage status]

This patch modifies the FullDetectorResponse code to more efficiently compute point-source responses, especially for the ScattMap case.  The main changes are:

 * vectorize rotation of PSRs from instrument to galactic coordinates
 * avoid allocating Histograms until their contents are fully computed

There are a few other minor cleanups to the response directory and one tweak to polarization angle to support slightly more vectorization, which is used by the new response code.

# BENCHMARKS

(1) Using the DC3 O3 continuum response and the first hour of orientation data from the 1s DC2/DC3 orientation file, call get_extended_source() with nside_image=8 (earth occlusion is enabled by default).

OLD CODE: 8233 s
NEW CODE: 820 s   (10x faster)

(2) Analogous to (1) using the DC3 polarization response (but we don't store the resulting PSRs because ExtendedSourceResponse does not support that).

OLD CODE: 72219 s
NEW CODE: 2405 s  (30x faster)

# PERFORMANCE NOTES

In my profiling of benchmark #1, the new code spends a bit under 25% of its time reading slices of the response from disk, a bit over 25% doing the math of rotating and adding PSRs, and much of the remaining time (at least 42%) doing coordinate transformations associated with HealpixAxis.find_bin() and Attitude construction.  The cost of coordinate transformations seems excessive, which is unfortunately a property of Astropy's SkyCoord class.

If this code should start using interpolation rather than find_bin in the future, perhaps we should look into doing rotations in the spherical harmonic domain instead.