@jdbuhler 2025-07-17
Missing lines in patch coverage are all error messages/exceptions